pyproject.toml
[project]
name = "aws-agentcore-runtime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai==1.3.0",
    "crewai-tools==1.3.0",
    "fastapi==0.121.0",
    "pip==25.3",
    "httpx==0.28.1",
    "pydantic==2.12.3",
    "strands-agents==1.14.0",
    "uvicorn[standard]==0.38.0",
    "litellm==1.79.1",
    "python-dotenv>=1.2.1",
]
